fix(build): replace uintptr_t with unsigned long; disable IRXJCL link
uintptr_t is not defined in GCC 3.2.3's stdint.h on the MVS target.
Replace both casts in irx#init.c with unsigned long, which is the
correct integer-to-pointer intermediate type on 24-bit/32-bit S/370.

Comment out the IRXJCL [[link.module]] block in project.toml.
IRXJCL CSECT does not exist yet; the link step was failing with RC=8
on unresolved lstring370 symbols.

+++ src/irx#init.c
@@ -228,7 +228,7 @@ int irx_init_initenvb(struct envblock *prev_envblock,
                 irxanchr_entry_t *slot = irx_anchor_find_by_tcb(tcb);
                 if (slot != NULL)
                 {
-                    prev = (struct envblock *)(uintptr_t)slot->envblock_ptr;
+                    prev = (struct envblock *)(unsigned long)slot->envblock_ptr;
                 }
             }
         }
@@ -330,7 +330,7 @@ int irx_init_initenvb(struct envblock *prev_envblock,
     memcpy(envblk->envblock_id, ENVBLOCK_ID, 8);
     memcpy(envblk->envblock_version, ENVBLOCK_VERSION_0042, 4);
     envblk->envblock_length = (int)sizeof(struct envblock);
-    envblk->envblock_userfield = (void *)(uintptr_t)user_field;
+    envblk->envblock_userfield = (void *)(unsigned long)user_field;
 
     /* ----------------------------------------------------------------
      * Step 5: PARMBLOCK copy allocation.
